requirements.txt
ollama
pymupdf
openai-whisper
chromadb
langchain-experimental
langchain-huggingface
